kundan/personal_post/pyproject.toml
[project]
name = "proj-copy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.37.0",
]
